requirements.txt
python-telegram-bot[job-queue]==21.7
flask==3.0.3
python-dotenv==1.0.1
